pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guitarduets"
version = "0.1.0"
description = "A codebase of Classical Guitar Duet Separation using GuitarDuets -- a Dataset of Real and Synthesized Guitar Recordings, Glytsos, Garoufis, Zlatintsi and Maragos"
requires-python = ">=3.10"
dependencies = [
  "dora-search>=0.1.12",
  "diffq>=0.2.1",
  "einops",
  "hydra-colorlog>=1.1",
  "hydra-core>=1.1",
  "julius>=0.2.3",
  "lameenc>=1.2",
  "museval",
  "openunmix",
  "PyYAML",
  "submitit",
  "torch>=1.8.1",
  "torchaudio>=0.8,<2.1",
  "tqdm",
  "treetable",
  "soundfile>=0.10.3;sys_platform=='win32'",
  "asteroid",
  "matplotlib",
  "mido",
  "numpy<2",
  "omegaconf",
  "pandas",
  "scipy",
  "threadpoolctl",
  "requests",
]

[tool.setuptools.packages.find]
where = ["src"]

